🐜 Fix template error blinkenlights
Because the change detection subscription is on the subject observables not the base observables, errors don't get caught in the merged observable if they aren't bound to the subjects any longer. On the one hand this provides somewhat better behavior: one observable failing is no longer a full template "stopped" event. On the other hand to fix this we need to rewire the error blinkenlights entirely, and in so doing absolutely must make some breaking changes, so SEMVER MAJOR bump here for what is ostensibly "a simple bug fix".